Add files via upload
First draft of the ER diagram and tables. Did not include cardinality yet because I figured there would still be a lot of changes. Will include that in the final version. Feel free to make any changes you want, I wasn't exactly sure what everyone wanted to do so this is just my first thought, just a place to start.